Fix bug where multiple group by exprs caused ambiguous column error
If you do GROUP BY `a`, `b`
where `a` and `b` are both aliases to the same column, Datafusion would group by both (unaliasing them), spit out a plan with two columns that are the same, and then have aliases which don't know which column to match to (since you have two with the same name, the original column). Silly datafusion!