planning: Fix issues with distinct on with aliases & fully qualified tables
DISTINCT ON and the ORDER BY can work with aliases given from the SELECT. In this case, you still need to understand if the DISTINCT ON and ORDER BY are referencing the same columns in order to correctly throw an error if not (as ORDER BY must have DISTINCT ON columns to begin with, but it doesn't matter if they're aliased or not).
This commit also fixes issues with fully qualified names not working in the comparison.